Added ncols to Model.
Now the model controls the number of columns. This has been used
for correcting a bug that happened when reading a file without field
names since an infinite list of names was created. Also, now when
updating a row to a position that is beyond its size, an appropriate
number of fields is created.