Created the PresenterAuto.
The PresenterAuto is an Auto that uses a WriterT IO so that
the writing of the Input and the GUICommands can be done
easily. To avoid circular imports, it has been necessary to
split the different files for the Input in two parts, one for
the data structure and one for the Auto.