Remove the backup on exit if the model is not changed.
When the model is not changed, the backup copy is removed.
Added a command ExitRequested that has the old meaning of
ExitProgram. Now, ExitProgram is used to force exit and
DoExit does the actual exit. ExitProgram is in charge of
removing the backup (via the new command RemoveBackup)
and then invoking DoExit.